+++ src/config.py
@@ -23,6 +23,4 @@ log_folder_path.mkdir(exist_ok=True)
 
 
 # GIT
-error_code_exceptions = {
-    1: "merge conflict"
-}
+error_code_exceptions = {1: "merge conflict"}