+++ src/logger.py
@@ -7,10 +7,6 @@ log = logging.getLogger("GTF")
 log.setLevel(log_level)
 formatter = logging.Formatter("%(asctime)s - %(name)s - %(levelname)s :: %(message)s")
 
-file_handler = TimedRotatingFileHandler(
-    filename=log_path,
-    when="W5",
-    interval=1
-)
+file_handler = TimedRotatingFileHandler(filename=log_path, when="W5", interval=1)
 file_handler.setFormatter(formatter)
 log.addHandler(file_handler)