+++ src/response.py
@@ -5,9 +5,10 @@ from src.exceptions import ResponseMismatchError
 
 
 class Response:
-    """ Class Response
+    """Class Response
     Contains the response segments from the git command.
     """
+
     def __init__(self, result):
         self.output = result.stdout
         self.output_list = result.stdout.splitlines()
@@ -17,16 +18,12 @@ class Response:
         if text in self.output:
             return True
 
-        raise ResponseMismatchError(
-            f"Response doesn't contain expected text: '{text}'.\n"
-            f"Response:\n\n{self.output}"
-        )
+        raise ResponseMismatchError(f"Response doesn't contain expected text: '{text}'.\nResponse:\n\n{self.output}")
 
     def has(self, on_line, text, regex=False):
         if len(self.output_list) < on_line or on_line < 0:
             raise IndexError(
-                f"Requested line number '{on_line}' "
-                f"is out of the response boundaries.\nResponse:\n\n{self.output}"
+                f"Requested line number '{on_line}' is out of the response boundaries.\nResponse:\n\n{self.output}"
             )
 
         line = self.output_list[on_line - 1]