+++ tests/test_branch.py
@@ -8,8 +8,8 @@ def test_create():
 
     scenario.run("git branch new-branch")
     response = scenario.run("git branch")
-    assert response.contains('* main')
-    assert response.contains('new-branch')
+    assert response.contains("* main")
+    assert response.contains("new-branch")
 
 
 def test_checkout():
@@ -18,21 +18,21 @@ def test_checkout():
 
     scenario.run("git checkout -b checkout-branch")
     response = scenario.run("git branch")
-    assert response.contains('main')
-    assert response.contains('* checkout-branch')
+    assert response.contains("main")
+    assert response.contains("* checkout-branch")
 
     response = scenario.run("git status")
-    assert response.contains('On branch checkout-branch')
-    assert response.contains('nothing to commit, working tree clean')
+    assert response.contains("On branch checkout-branch")
+    assert response.contains("nothing to commit, working tree clean")
 
     scenario.run("git checkout main")
     response = scenario.run("git branch")
-    assert response.contains('* main')
-    assert response.contains('checkout-branch')
+    assert response.contains("* main")
+    assert response.contains("checkout-branch")
 
     response = scenario.run("git status")
-    assert response.contains('On branch main')
-    assert response.contains('nothing to commit, working tree clean')
+    assert response.contains("On branch main")
+    assert response.contains("nothing to commit, working tree clean")
 
 
 def test_merge():
@@ -61,36 +61,36 @@ def test_merge():
     scenario.run(f"git merge {merge_branch}")
 
     response = scenario.run("git status")
-    assert response.contains('On branch main')
-    assert response.contains('nothing to commit, working tree clean')
+    assert response.contains("On branch main")
+    assert response.contains("nothing to commit, working tree clean")
 
     response = scenario.run("git log")
-    assert response.has(on_line=1, text=r'commit [0-9a-f]{40}', regex=True)
-    assert response.has(on_line=2, text=r'Merge: [0-9a-f]{7} [0-9a-f]{7}', regex=True)
-    assert response.has(on_line=3, text='Author: Pavle Mutic <mail@pavlemutic.com>')
-    assert response.has(on_line=4, text=r'Date:   \w{3} \w{3} \d{1,2} \d{2}:\d{2}:\d{2} \d{4} [+-]\d{4}', regex=True)
+    assert response.has(on_line=1, text=r"commit [0-9a-f]{40}", regex=True)
+    assert response.has(on_line=2, text=r"Merge: [0-9a-f]{7} [0-9a-f]{7}", regex=True)
+    assert response.has(on_line=3, text="Author: Pavle Mutic <mail@pavlemutic.com>")
+    assert response.has(on_line=4, text=r"Date:   \w{3} \w{3} \d{1,2} \d{2}:\d{2}:\d{2} \d{4} [+-]\d{4}", regex=True)
     assert response.has(on_line=6, text="Merge branch 'branch-to-merge'")
 
-    assert response.has(on_line=8, text=r'commit [0-9a-f]{40}', regex=True)
-    assert response.has(on_line=9, text='Author: Pavle Mutic <mail@pavlemutic.com>')
-    assert response.has(on_line=10, text=r'Date:   \w{3} \w{3} \d{1,2} \d{2}:\d{2}:\d{2} \d{4} [+-]\d{4}', regex=True)
+    assert response.has(on_line=8, text=r"commit [0-9a-f]{40}", regex=True)
+    assert response.has(on_line=9, text="Author: Pavle Mutic <mail@pavlemutic.com>")
+    assert response.has(on_line=10, text=r"Date:   \w{3} \w{3} \d{1,2} \d{2}:\d{2}:\d{2} \d{4} [+-]\d{4}", regex=True)
     assert response.has(on_line=12, text='"Add sixth line to merge_file"')
 
-    assert response.has(on_line=14, text=r'commit [0-9a-f]{40}', regex=True)
-    assert response.has(on_line=15, text='Author: Pavle Mutic <mail@pavlemutic.com>')
-    assert response.has(on_line=16, text=r'Date:   \w{3} \w{3} \d{1,2} \d{2}:\d{2}:\d{2} \d{4} [+-]\d{4}', regex=True)
+    assert response.has(on_line=14, text=r"commit [0-9a-f]{40}", regex=True)
+    assert response.has(on_line=15, text="Author: Pavle Mutic <mail@pavlemutic.com>")
+    assert response.has(on_line=16, text=r"Date:   \w{3} \w{3} \d{1,2} \d{2}:\d{2}:\d{2} \d{4} [+-]\d{4}", regex=True)
     assert response.has(on_line=18, text='"Update second line in merge_file"')
 
-    assert response.has(on_line=20, text=r'commit [0-9a-f]{40}', regex=True)
-    assert response.has(on_line=21, text='Author: Pavle Mutic <mail@pavlemutic.com>')
-    assert response.has(on_line=22, text=r'Date:   \w{3} \w{3} \d{1,2} \d{2}:\d{2}:\d{2} \d{4} [+-]\d{4}', regex=True)
+    assert response.has(on_line=20, text=r"commit [0-9a-f]{40}", regex=True)
+    assert response.has(on_line=21, text="Author: Pavle Mutic <mail@pavlemutic.com>")
+    assert response.has(on_line=22, text=r"Date:   \w{3} \w{3} \d{1,2} \d{2}:\d{2}:\d{2} \d{4} [+-]\d{4}", regex=True)
     assert response.has(on_line=24, text='"Add merge_file"')
 
-    assert response.has(on_line=26, text=r'commit [0-9a-f]{40}', regex=True)
-    assert response.has(on_line=27, text='Author: Pavle Mutic <mail@pavlemutic.com>')
-    assert response.has(on_line=28, text=r'Date:   \w{3} \w{3} \d{1,2} \d{2}:\d{2}:\d{2} \d{4} [+-]\d{4}', regex=True)
+    assert response.has(on_line=26, text=r"commit [0-9a-f]{40}", regex=True)
+    assert response.has(on_line=27, text="Author: Pavle Mutic <mail@pavlemutic.com>")
+    assert response.has(on_line=28, text=r"Date:   \w{3} \w{3} \d{1,2} \d{2}:\d{2}:\d{2} \d{4} [+-]\d{4}", regex=True)
 
-    assert response.has(on_line=30, text='Initial commit')
+    assert response.has(on_line=30, text="Initial commit")
 
     assert scenario.get_file(merge_file) == scenario.get_expected_file(expected_merge_file)
 
@@ -119,34 +119,34 @@ def test_merge_conflict():
 
     scenario.run(f"git checkout main")
     response = scenario.run(f"git merge {branch}")
-    assert response.has(on_line=1, text='Auto-merging conflict_file')
-    assert response.has(on_line=2, text='CONFLICT (content): Merge conflict in conflict_file')
-    assert response.has(on_line=3, text='Automatic merge failed; fix conflicts and then commit the result.')
+    assert response.has(on_line=1, text="Auto-merging conflict_file")
+    assert response.has(on_line=2, text="CONFLICT (content): Merge conflict in conflict_file")
+    assert response.has(on_line=3, text="Automatic merge failed; fix conflicts and then commit the result.")
 
     response = scenario.run("git status")
-    assert response.has(on_line=1, text='On branch main')
-    assert response.has(on_line=2, text='You have unmerged paths.')
+    assert response.has(on_line=1, text="On branch main")
+    assert response.has(on_line=2, text="You have unmerged paths.")
     assert response.has(on_line=3, text='(fix conflicts and run "git commit")')
     assert response.has(on_line=4, text='(use "git merge --abort" to abort the merge)')
-    assert response.has(on_line=6, text='Unmerged paths:')
+    assert response.has(on_line=6, text="Unmerged paths:")
     assert response.has(on_line=7, text='(use "git add <file>..." to mark resolution)')
-    assert response.has(on_line=8, text='both modified:   conflict_file')
+    assert response.has(on_line=8, text="both modified:   conflict_file")
     assert response.has(on_line=10, text='no changes added to commit (use "git add" and/or "git commit -a")')
 
     response = scenario.run("git log")
-    assert response.has(on_line=1, text=r'commit [0-9a-f]{40}', regex=True)
-    assert response.has(on_line=2, text='Author: Pavle Mutic <mail@pavlemutic.com>')
-    assert response.has(on_line=3, text=r'Date:   \w{3} \w{3} \d{1,2} \d{2}:\d{2}:\d{2} \d{4} [+-]\d{4}', regex=True)
+    assert response.has(on_line=1, text=r"commit [0-9a-f]{40}", regex=True)
+    assert response.has(on_line=2, text="Author: Pavle Mutic <mail@pavlemutic.com>")
+    assert response.has(on_line=3, text=r"Date:   \w{3} \w{3} \d{1,2} \d{2}:\d{2}:\d{2} \d{4} [+-]\d{4}", regex=True)
     assert response.has(on_line=5, text='"Update conflict_file on 4th from main"')
 
-    assert response.has(on_line=7, text=r'commit [0-9a-f]{40}', regex=True)
-    assert response.has(on_line=8, text='Author: Pavle Mutic <mail@pavlemutic.com>')
-    assert response.has(on_line=9, text=r'Date:   \w{3} \w{3} \d{1,2} \d{2}:\d{2}:\d{2} \d{4} [+-]\d{4}', regex=True)
+    assert response.has(on_line=7, text=r"commit [0-9a-f]{40}", regex=True)
+    assert response.has(on_line=8, text="Author: Pavle Mutic <mail@pavlemutic.com>")
+    assert response.has(on_line=9, text=r"Date:   \w{3} \w{3} \d{1,2} \d{2}:\d{2}:\d{2} \d{4} [+-]\d{4}", regex=True)
     assert response.has(on_line=11, text='"Add conflict_file"')
 
-    assert response.has(on_line=13, text=r'commit [0-9a-f]{40}', regex=True)
-    assert response.has(on_line=14, text='Author: Pavle Mutic <mail@pavlemutic.com>')
-    assert response.has(on_line=15, text=r'Date:   \w{3} \w{3} \d{1,2} \d{2}:\d{2}:\d{2} \d{4} [+-]\d{4}', regex=True)
-    assert response.has(on_line=17, text='Initial commit')
+    assert response.has(on_line=13, text=r"commit [0-9a-f]{40}", regex=True)
+    assert response.has(on_line=14, text="Author: Pavle Mutic <mail@pavlemutic.com>")
+    assert response.has(on_line=15, text=r"Date:   \w{3} \w{3} \d{1,2} \d{2}:\d{2}:\d{2} \d{4} [+-]\d{4}", regex=True)
+    assert response.has(on_line=17, text="Initial commit")
 
     assert scenario.get_file(file) == scenario.get_expected_file(expected_conflict_file)