+++ unit/test_scenario.py
@@ -6,6 +6,7 @@ from src.response import Response
 from src.config import output_path, files_path, expected_files_path
 from src.exceptions import GitExecutionError
 
+
 @pytest.fixture
 def scenario():
     return Scenario(name="test.scenario", remote=True)
@@ -15,12 +16,10 @@ def scenario():
 def test_add_file(mock_copyfile, scenario):
     scenario.scenario_local_path = output_path / "test_scenario" / "local"
     scenario.add_file("source.txt", "dest.txt")
-    mock_copyfile.assert_called_once_with(
-        src=files_path / "source.txt",
-        dst=scenario.scenario_local_path / "dest.txt"
-    )
+    mock_copyfile.assert_called_once_with(src=files_path / "source.txt", dst=scenario.scenario_local_path / "dest.txt")
     assert "dest.txt" in scenario._files
     assert isinstance(scenario._files["dest.txt"], File)
+
 
 def test_get_file(scenario):
     file_mock = MagicMock(spec=File)
@@ -28,11 +27,13 @@ def test_get_file(scenario):
     assert scenario.get_file("test.txt") == file_mock
     assert scenario.get_file("nonexistent.txt") is None
 
+
 def test_get_expected_file():
     file_name = "expected.txt"
     expected_file = Scenario.get_expected_file(file_name)
     assert isinstance(expected_file, File)
     assert expected_file.path == expected_files_path / file_name
+
 
 @patch("src.scenario.subprocess.run")
 def test_run_success(mock_run, scenario):
@@ -44,9 +45,21 @@ def test_run_success(mock_run, scenario):
     )
     assert isinstance(response, Response)
 
+
 @patch("src.scenario.subprocess.run")
 def test_run_failure(mock_run, scenario):
+    mock_run.return_value = MagicMock(returncode=2, stderr="Error output")
+    command = "git status"
+    with pytest.raises(GitExecutionError, match="Git command 'git status' failed with status code '2'"):
+        scenario.run(command)
+
+
+@patch("src.scenario.subprocess.run")
+def test_run_failure_exception(mock_run, scenario):
     mock_run.return_value = MagicMock(returncode=1, stderr="Error output")
     command = "git status"
-    with pytest.raises(GitExecutionError, match="Git command 'git status' failed with status code '1'"):
-        scenario.run(command)
+    response = scenario.run(command)
+    mock_run.assert_called_once_with(
+        ["git", "status"], cwd=scenario.scenario_local_path, capture_output=True, text=True
+    )
+    assert isinstance(response, Response)